requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
langchain==0.0.346
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3
faiss-gpu==1.7.4